UserActions history refactor
