+++ app/state/video_debate/reducer.js
@@ -3,13 +3,11 @@ import { combineReducers } from "redux"
 import VideoReducer from './video/reducer'
 import CommentsReducer from './comments/reducer'
 import StatementsReducer  from './statements/reducer'
-import VideoDebateHistoryReducer  from './history/reducer'
 
 
 const VideoDebateReducer = combineReducers({
   video: VideoReducer,
   statements: StatementsReducer,
-  comments: CommentsReducer,
-  history: VideoDebateHistoryReducer
+  comments: CommentsReducer
 })
 export default VideoDebateReducer
