+++ app/styles/_components/UsersActions/entity.sass
@@ -1,0 +1,10 @@
+.entity
+  .title .statement-text
+    display: inline-block
+
+  .speaker-preview
+    margin: 0 auto
+    width: fit-content
+    border: 1px solid $grey-lighter
+    border-radius: 10px
+    padding: 1em 1.5em
