+++ app/styles/_components/UsersActions/history.sass
@@ -1,0 +1,19 @@
+.user-actions-history
+  @extend .box
+  background: none
+  box-shadow: none
+  border-radius: 0
+  margin: 0 auto
+  width: 90%
+  max-width: 900px
+  padding-bottom: 1rem
+  .expend-old-actions
+    display: block
+    width: 100%
+    padding-top: 10px
+  .old-actions
+    padding: 0 25px
+    .separator
+      width: 50%
+      height: 25px
+      border-right: 1px solid #eaeaea
