+++ app/styles/_components/UsersActions/user_action.sass
@@ -1,0 +1,45 @@
+.user-action
+  .action-description
+    text-align: left
+    padding: 15px
+    &> *
+      margin-right: 3px
+    .action-type
+      margin-right: 7px
+  .action-buttons
+    &> a
+      @extend .card-footer-item
+      padding: 5px
+      background-color: whitesmoke
+  .action-diff
+    padding: 15px
+    text-align: left
+    box-shadow: 0px 2px 7px 0px #212121 inset
+    background-color: $dark
+    transition: height 0.4s
+    .diff-entry
+      &:not(:last-child)
+        margin-bottom: 5px
+      .diff-key
+        display: inline-block
+        margin-right: 5px
+        font-weight: bold
+        min-width: 80px
+        color: orange
+        vertical-align: top
+      .diff-view
+        display: inline-block
+        background: none
+        font-size: 1em
+        white-space: pre-wrap
+        span
+          color: $dark-invert
+          &.added, &.removed
+            padding: 2px
+          &.added
+            background-color: #006330
+          &.removed
+            white-space: normal
+            background-color: darkred
+            text-decoration: line-through
+            opacity: 0.45
