+++ app/styles/_components/modal.sass
@@ -29,13 +29,6 @@ $modal-background: rgba(10, 10, 10, 0.50)
     .modal-card-body
       padding: 35px 20px
 
-.modal-history
-  .modal-card
-    width: 800px
-    .modal-card-body
-      text-align: center
-      padding: 0
-
 .form-buttons
   text-align: center
   width: 100%
