+++ app/styles/application.sass
@@ -50,6 +50,9 @@
 @import "_components/User/picture"
 @import "_components/User/user_page"
 @import "_components/User/user_reset_password"
+@import "_components/UsersActions/entity"
+@import "_components/UsersActions/history"
+@import "_components/UsersActions/user_action"
 @import "_components/Utils/error_page"
 @import "_components/Utils/loading_frame"
 @import "_components/Utils/popover"
