+++ package.json
@@ -24,6 +24,7 @@
     "isomorphic-fetch": "^2.2.1",
     "marked": "^0.3.6",
     "phoenix": "^1.3.0",
+    "prop-types": "^15.6.0",
     "re-reselect": "^1.0.1",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
